Refactor icon locations
