+++ frontend/src/images/index.js
@@ -1,0 +1,6 @@
+export { ReactComponent as ArrowIcon } from './arrow.svg';
+export { ReactComponent as CafecitoIcon } from './cafecito.svg';
+export { ReactComponent as DeleteIcon } from './delete.svg';
+export { ReactComponent as GoogleIcon } from './google.svg';
+export { ReactComponent as InstagramIcon } from './instagram.svg';
+export { ReactComponent as NoConnectionIcon } from './noConnection.svg';
